Simplify dirs and make Route available in higher namespace
