+++ lib/travis/yaml/web/route.rb
@@ -1,15 +1,11 @@
-module Travis::Yaml::Web::V1
-  module Routes
-    module Route
-      def call(env)
-        public_send(env['REQUEST_METHOD'.freeze].downcase, env)
-      end
+module Travis::Yaml::Web::Route
+  def call(env)
+    public_send(env['REQUEST_METHOD'.freeze].downcase, env)
+  end
 
-      private
+  private
 
-      def method_missing(*)
-        [404, {}, []]
-      end
-    end
+  def method_missing(*)
+    [404, {}, []]
   end
 end
