+++ lib/travis/yaml/web/v1/parse.rb
@@ -1,0 +1,25 @@
+require 'oj'
+require 'travis/yaml'
+require 'travis/yaml/web/route'
+
+module Travis::Yaml::Web
+  module V1
+    class Parse
+      include Route
+
+      def post(env)
+        req = Rack::Request.new(env)
+        body = req.body.read
+        config = Travis::Yaml.load(body)
+
+        response = {
+          'version' => 'v1',
+          'messages' => config.msgs.map { |m| Travis::Yaml.msg(m) },
+          'config' => config.serialize
+        }
+
+        [200, { 'Content-Type' => 'application/json' }, [Oj.dump(response)]]
+      end
+    end
+  end
+end
